secured admin area for adding , editing deleting products to the site owner only

+++ products/views.py
@@ -1,6 +1,7 @@
 """ Views imports for products"""
 from django.shortcuts import render, get_object_or_404, reverse, redirect
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Product, Category
@@ -78,10 +79,15 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     """
     Adding products to the store
     """
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Only the store owner can complete this task')
+        return redirect(reverse('home'))
 
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
@@ -102,10 +108,16 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """
     Edit a product
     """
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Only the store owner can complete this task')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -114,7 +126,7 @@ def edit_product(request, product_id):
             messages.success(request, 'Successfully updated product')
             return redirect(reverse('product_detail', args=[product.id]))
         else:
-            messages.error(request, 'Faild to update, please check input')
+            messages.error(request, 'Failed to update, please check input')
     else:
 
         form = ProductForm(instance=product)
@@ -129,10 +141,15 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """
     Delete a product
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Only the store owner can complete this task')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, f'You have deleted {product.name}')
